+++ src/tests/validator.test.js
@@ -1,19 +1,20 @@
-import Validator, { rules } from '../'
+/* global describe,expect,test */
+import Validator, { rules } from '..'
 
 describe('test validate', () => {
   Validator.setMessages({
     required: 'item required'
   })
 
-  const form = {
-    username: "test",
+  const dataForm = {
+    username: 'test',
     password: '',
     retype: 'test',
     bio: 'test',
     age: 15,
     nickanme: 'tEst',
     tags: ['hello', 'world']
-  };
+  }
   const validationConfig = {
     username: [
       {
@@ -26,23 +27,21 @@ describe('test validate', () => {
       rules.REQUIRED,
       {
         ...rules.LENGTH_BETWEEN,
-        value: [6, 20],
+        value: [6, 20]
       }
     ],
     retype: [
       rules.REQUIRED,
       {
         ...rules.FUNCTION,
-        value: (value, rule, form) => {
-          return value === form.password
-        },
+        value: (value, rule, form) => value === form.password,
         message: 'retype error'
       }
     ],
     bio: [
       {
         ...rules.MAX_LENGTH,
-        value: 3,
+        value: 3
       }, {
         ...rules.PATTERN,
         value: /[a-zA-Z]+/
@@ -70,23 +69,79 @@ describe('test validate', () => {
   }
 
   test('should return false', () => {
-    const { result, messages } = Validator.validate(form, validationConfig);
+    const { result, messages } = Validator.validate(dataForm, validationConfig)
 
-    expect(result).toBeFalsy();
+    expect(result).toBeFalsy()
 
     expect(messages).toEqual({
-      "password": ['item required',"length btween 6 to 20 required"],
-      "retype": [ "retype error"],
-      "username": ["invalid, lenth less than 6 charactors"],
-      "age": [
-        "number btween 18 to 60 required",
+      password: ['item required', 'length btween 6 to 20 required'],
+      retype: ['retype error'],
+      username: ['invalid, lenth less than 6 charactors'],
+      age: [
+        'number btween 18 to 60 required'
       ],
-      "bio": [
-        "less than 3 chars required",
+      bio: [
+        'less than 3 chars required'
       ],
-      "nickname": [
-        "format not matched",
+      nickname: [
+        'format not matched'
+      ]
+    })
+  })
+
+  test('nested rules', () => {
+    const validateForm = {
+      tagId: 'test',
+      login: {
+        username: 'test',
+        password: '',
+        retype: 'test',
+        info: {
+          age: 10
+        }
+      }
+    }
+
+    const validateRules = {
+      tagId: [
+        rules.NUMBER
       ],
+      login: {
+        username: [
+          rules.REQUIRED,
+          {
+            ...rules.MIN_LENGTH,
+            value: 6
+          }
+        ],
+        password: [
+          rules.REQUIRED
+        ],
+        retype: [{
+          ...rules.FUNCTION,
+          value: (value, rule, form) => value === form.password
+        }],
+        info: {
+          age: [{
+            ...rules.MIN,
+            value: 18
+          }]
+        }
+      }
+    }
+
+    const { result, messages } = Validator.validate(validateForm, validateRules)
+
+    expect(result).toBeFalsy()
+
+    expect(messages).toEqual({
+      login: {
+        info: { age: ['more than 18 required'] },
+        password: ['item required'],
+        retype: ['invalid'],
+        username: ['more than 6 chars required']
+      },
+      tagId: ['number required']
     })
   })
 })