+++ CHANGELOG.md
@@ -1,6 +1,7 @@
 ## 1.0.4 (Jue 28, 2019)
 
 1. fix incorrect implementation of validate required rule
+2. add validate nested rules
 
 ## 1.0.3 (Jue 24, 2019)